feat 9: useGif custom hook
