+++ src/GifsApp.tsx
@@ -1,40 +1,13 @@
-import { useState } from 'react'
 import { GifList } from './gifs/GifList'
 import { PreviousSearches } from './gifs/PreviousSearches'
 import { CustomHeader } from './shared/components/CustomHeader'
 import { SearchBar } from './shared/components/SearchBar'
 import { mockGifs } from './mock-data/gifs.mock'
 
-import { getGifsByQuery } from './gifs/actions/get-gifs-by-query.action'
-import type { Gif } from './gifs/interfaces/gif-interface'
+import { useGifs } from './gifs/hooks/useGifs'
 
 export const GifsApp = () => {
-  const [gifs, setGifs] = useState<Gif[]>([...mockGifs])
-  const [previousTerms, setPreviousTerms] = useState<string[]>([])
-
-  const handleTermClicked = (term: string) => {
-    console.log({ term })
-  }
-
-  const handleSearch = async (query: string) => {
-    query = query.trim().toLowerCase()
-
-    if (query === '') return
-
-    // console.log({ query })
-
-    if (!previousTerms.includes(query)) {
-      // Add the new term to the beginning of the list
-      // Keep only the last 8 terms
-      setPreviousTerms((prev) => [query, ...prev].splice(0, 8))
-
-      const gifs = await getGifsByQuery(query)
-
-      console.log({ gifs })
-
-      setGifs(gifs)
-    }
-  }
+  const { gifs, previousTerms, handleTermClicked, handleSearch } = useGifs([...mockGifs], [])
 
   return (
     <>
@@ -50,7 +23,7 @@ export const GifsApp = () => {
       {/* Gifs */}
       <GifList gifs={gifs} />
 
-      <span> {JSON.stringify(gifs)}</span>
+      {/* <span> {JSON.stringify(gifs)}</span> */}
     </>
   )
 }
